Eliminate the following warnings from pylint when run on all *.py files: C0200: Consider using enumerate instead of iterating with range and len (consider-using-enumerate) C0301: Line too long W0107: Unnecessary pass statement (unnecessary-pass) W0109: Duplicate key 'source_ref_text' in dictionary (duplicate-key) W0612: Unused variable W0613: Unused argument
Eliminate all but a few warnings from flake8 by breaking up lines to shorten them.  The remaining too-long lines are made too long by very long field names, and will be hard to shorten without rrenaming.

Stop using '_' as an alias for ugettext, so that it can be used pythonically as a to-be-ignored parameter name.